Migrate from packer to lazy.nvim
Had to let lazy manage treesitter plugins, as lazy and nix are currently
incompatible since lazy resets the packpath and rwp. Disabling this
reset leads to slower startuptime, so it's not worth it.

I changed to let nix manage treesitter parsers due to updates happening
left and right. Treesitter seems to have added a lockfile, so updates
should now happen only when treesitter is updated.