fish: fix issue with universal variables
-U adds the config to ~/.config/fish/fish_variables. If a change is made
there it cannot be undone with nix, which makes the config not
declarative.

It also caused issues with $EDITOR. Macos sets it to nano, which is then
inherited by the fish shell when it starts from the terminal emulator.
Since universal variables are loaded last, $EDITOR was never set to
neovim.